+++ wicket/src/cli/preflight.rs
