+++ wicket/src/cli/rack_setup.rs
