+++ wicket/src/cli/rack_setup/config_template.toml
